provisioning-station/requirements.txt
pyserial>=3.5
requests>=2.31.0
PyJWT>=2.8
